src/requirements.txt
djangorestframework==3.12.*
markdown==3.3.*
psycopg2-binary==2.9.*
djoser==2.1.*
djangorestframework_simplejwt==4.8.*
django-cors-headers==3.8.*
drf-yasg==1.20.*
drf-nested-routers==0.93.*